Add nav bar and react router
