+++ e-commerce-website/src/App.js
@@ -1,38 +1,24 @@
-import "./categories.styles.scss";
-import CategoryDirectory from "./components/category-directory/category-directory.component";
+import Home from './routes/home/home.component.jsx'
+import { Routes, Route} from 'react-router-dom'
+import Navigation from './routes/navigation/navigation.component.jsx';
+
+
+const Shop = () => {
+  return (
+    <h1>I am the shop page</h1>
+  )
+}
 
 const App = () => {
-   const categories = [
-     {
-       id: 1,
-       title: "hats",
-       imageUrl: "https://i.ibb.co/cvpntL1/hats.png",
-     },
-     {
-       id: 2,
-       title: "jackets",
-       imageUrl: "https://i.ibb.co/px2tCc3/jackets.png",
-     },
-     {
-       id: 3,
-       title: "sneakers",
-       imageUrl: "https://i.ibb.co/0jqHpnp/sneakers.png",
-     },
-     {
-       id: 4,
-       title: "womens",
-       imageUrl: "https://i.ibb.co/GCCdy8t/womens.png",
-     },
-     {
-       id: 5,
-       title: "mens",
-       imageUrl: "https://i.ibb.co/R70vBrQ/men.png",
-     },
-   ];
+  
   return (
-    <div>
-      <CategoryDirectory categories = {categories}/>
-    </div>
+    <Routes>
+      <Route path='/' element= {<Navigation/>}>
+        <Route path='shop' element= {<Shop />}></Route>
+      <Route index element= {<Home/>} />
+      </Route>
+    </Routes>
+    
   );
 };
